requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.20.0
opencv-python>=4.5.0
pillow>=8.0.0
tqdm>=4.60.0
pandas>=1.3.0
zipfile36>=0.1.0
git+https://github.com/Parskatt/RoMa.git
